@@ -13,17 +13,17 @@ final class Container implements ContainerInterface
     /**
      * @var array<string, callable>
      */
-    private $factories = [];
+    private array $factories = [];
 
     /**
      * @var array<string, callable>
      */
-    private $prototypeFactories = [];
+    private array $prototypeFactories = [];
 
     /**
      * @var array<string, mixed>
      */
-    private $services = [];
+    private array $services = [];
 
     /**
      * @param array<string, callable> $factories
